move remoteed25519 to remote.rs
