+++ identity-account/src/crypto/remote.rs
@@ -3,12 +3,54 @@
 
 use core::marker::PhantomData;
 
+use serde::Serialize;
+
+use identity_core::convert::ToJson;
+use identity_core::crypto::Named;
+use identity_core::crypto::SetSignature;
+use identity_core::crypto::Signature;
+use identity_core::crypto::SignatureValue;
 use identity_core::error::Error;
 use identity_core::error::Result;
+use identity_core::utils::encode_b58;
 
 use crate::identity::IdentityId;
 use crate::storage::Storage;
 use crate::types::KeyLocation;
+
+pub struct RemoteEd25519;
+
+impl Named for RemoteEd25519 {
+  const NAME: &'static str = "JcsEd25519Signature2020";
+}
+
+impl RemoteEd25519 {
+  pub async fn create_signature<U, T>(data: &mut U, method: impl Into<String>, secret: &RemoteKey<'_, T>) -> Result<()>
+  where
+    U: Serialize + SetSignature,
+    T: Storage,
+  {
+    data.set_signature(Signature::new(Self::NAME, method));
+
+    let value: SignatureValue = Self::sign(&data, secret).await?;
+    let write: &mut Signature = data.try_signature_mut()?;
+
+    write.set_value(value);
+
+    Ok(())
+  }
+
+  pub async fn sign<T, X>(data: &X, remote_key: &RemoteKey<'_, T>) -> Result<SignatureValue>
+  where
+    X: Serialize,
+    T: Storage,
+  {
+    let message: Vec<u8> = data.to_jcs()?;
+    let signature: Vec<u8> = RemoteSign::sign(&message, remote_key).await?;
+    let signature: String = encode_b58(&signature);
+    Ok(SignatureValue::Signature(signature))
+  }
+}
 
 /// A reference to a storage instance and identity key location.
 #[derive(Debug)]
