+++ identity-account/src/tests/commands.rs
@@ -446,40 +446,3 @@ async fn test_delete_method() -> Result<()> {
 
   Ok(())
 }
-#[tokio::test]
-async fn test_account_shared_ref() -> Result<()> {
-  let stronghold_path: PathBuf = "./example-strong.hodl".into();
-
-  let storage = Stronghold::new(&stronghold_path, Some("my-password")).await?;
-  let account: Arc<Account> = Arc::new(Account::with_config(storage, Config::new().testmode(true)).await?);
-
-  let identity1 = account.create_identity(IdentityCreate::default()).await?;
-  let iota_did1: IotaDID = identity1.try_did()?.to_owned();
-
-  let identity2 = account.create_identity(IdentityCreate::default()).await?;
-  let iota_did2: IotaDID = identity2.try_did()?.to_owned();
-
-  let account_clone = Arc::clone(&account);
-
-  let task = tokio::task::spawn(async move {
-    update(account_clone.as_ref(), iota_did1).await;
-  });
-
-  update(account.as_ref(), iota_did2).await;
-
-  task.await.unwrap();
-
-  Ok(())
-}
-
-async fn update(account: &Account, did: IotaDID) {
-  for i in 0..2 {
-    account
-      .update_identity(&did)
-      .create_method()
-      .fragment(format!("key-{}", i))
-      .apply()
-      .await
-      .unwrap();
-  }
-}
